Move `store` module to `hash_utils` crate and add more docs to utils
