+++ compiler/hash-typecheck/src/exhaustiveness/mod.rs
@@ -40,7 +40,7 @@
 //!
 //! The implementation of this algorithm is based on the research paper:
 //! http://moscova.inria.fr/~maranget/papers/warn/warn.pdf, and is heavily
-//! inspired by the Rust Compiler implementation:
+//! inspired by and based off of the Rust Compiler implementation:
 //! <https://github.com/rust-lang/rust/tree/master/compiler/rustc_mir_build/src/thir/pattern/usefulness.rs>
 #![allow(dead_code)]
 use std::fmt::Display;
